End-to-end testing working perfectly.
At this stage, the express.js app and the mongo database are running as two separate services and are working perfectly based on the end-to-end tests I created in SIT 331. The express.js app is capable of connecting to the database that exists on a different container via using its service name (Docker has a DNS lookup system to my understanding).