Authentication microservice updated
At this stage, I have implemented a service that can handle communicating with the users microservice for the user data. This allows the authentication route to work. However, due to the session logic I've implemented, I'm only treated as authenticated in the authentication microservice and not the other microservices. This will need further work.